// src/enums/order-action.ts
export enum OrderAction {
  SUBMIT_CONTACT_FORM = 'submitContactForm',
  RESET_ORDER_PROCESS = 'resetOrderProcess',
  SUBMIT_DELIVERY_FORM = 'submitDeliveryForm',
  SUBMIT_PAYMENT_FORM = 'submitPaymentForm',
  TOGGLE_EDITABLE_FORM = 'toggleEditableForm',
  CONFIRM_ORDER = 'confirmOrder',
  TOGGLE_RULES = 'toggleRules',
  ORDER_READY = 'orderReady',
  EDIT_FORM = 'editForm',
}